fix: `WebServer#run` should return after we send `WebServer#stop`
Because `FTW::Server#each_connection` runs in a `while true` loop, breaking
the current iteration when we encounter a message on our stop pipe is not
enough; it allows the thread sending `FTW::Server#stop` to acquire the lock
and clean up our bound sockets, but the loop continues indefinitely.

Instead, only run the loop if we still have sockets to listen on.